+++ mobile/lib/services/filedata/model/file_data.dart
@@ -1,4 +1,5 @@
 import "package:photos/models/ml/face/face.dart";
+import "package:photos/utils/parse.dart";
 
 const _faceKey = 'face';
 const _clipKey = 'clip';
